requirements.txt
pandas
numpy
scikit-learn
matplotlib
nltk
seaborn
psycopg2-binary
sqlalchemy
mysql-connector-python
dill